Remove enable-by-serial-no for t265 device
A bug persists in the librealsense2 library that is preventing
the t265 device from being enabled via serial number.
https://github.com/IntelRealSense/librealsense/issues/5614

This is a temporary workaround until that bug is addressed. Note that multiple
t265 devices will not be able to be run at this time. The only other
workaround I can think of is reimplementing the stream configuration
with sensor API instead of pipeline API but that would be a major
refactor.